+++ tools/mcrun/make.esp.mk
@@ -76,13 +76,16 @@ endif
 all: $(LAUNCH)
 	
 debug: $(ARCHIVE)
-	$(shell pkill serial2xsbug)
-	$(START_XSBUG)
-	$(SERIAL2XSBUG) $(UPLOAD_PORT) $(DEBUGGER_SPEED) 8N1 -install $(ARCHIVE)
+	ifneq ($(BUILD_ONLY),1)
+		$(shell pkill serial2xsbug)
+		$(START_XSBUG)
+		$(SERIAL2XSBUG) $(UPLOAD_PORT) $(DEBUGGER_SPEED) 8N1 -install $(ARCHIVE)
+	endif
 
 release: $(ARCHIVE)
-	$(shell pkill serial2xsbug)
-	$(SERIAL2XSBUG) $(UPLOAD_PORT) $(DEBUGGER_SPEED) 8N1 -install $(ARCHIVE)
+		$(shell pkill serial2xsbug)
+		$(SERIAL2XSBUG) $(UPLOAD_PORT) $(DEBUGGER_SPEED) 8N1 -install $(ARCHIVE)
+	endif
 
 debugURL: $(ARCHIVE)
 	@echo "# curl "$(NAME)".xsa "$(URL)
